--- a/dlk/python/dlk/templates/src/func/conv2d.cpp
+++ b/dlk/python/dlk/templates/src/func/conv2d.cpp
@@ -45,16 +45,13 @@ void conv3x3_kn2row(T input[],
   assert(kh == 3 && kw == 3);
   assert(MAX_SIZE_KN2ROW_BUFFER_PER_LAYER >= oc * kh * kw * ih * iw);
 
-  // need to initialize output
-  std::memset(output, 0, oc * ih * iw * sizeof(U));
-
   Measurement::Start("kn2row");
   Measurement::Start("kn2row-buf");
 
   assert(p.input_height > 0);
   assert(p.input_width > 0);
 
-  U *buf = new U[MAX_SIZE_KN2ROW_BUFFER_PER_LAYER]();
+  static U buf[MAX_SIZE_KN2ROW_BUFFER_PER_LAYER];
 
   Measurement::Stop();
 
@@ -65,7 +62,6 @@ void conv3x3_kn2row(T input[],
 
   dlk::matrix_multiplication(kernels_, input_, buf_);
   dlk::matrix_shift_add(buf_, output_, p);
-  delete[] buf;
 
   Measurement::Stop();
 }
@@ -103,7 +99,7 @@ void conv1x1_kn2row(T input[],
 
    assert(p.input_height > 0);
    assert(p.input_width > 0);
-
+  
    auto kernels_ = dlk::MatrixView<T, dlk::MatrixOrder::RowMajor>(kernels, oc * kh * kw, ic);
    auto input_ = dlk::MatrixView<T, dlk::MatrixOrder::ColMajor>(input, ic, p.input_height * p.input_width);
    auto output_ = dlk::MatrixView<U, dlk::MatrixOrder::ColMajor>(output, oc, p.input_height * p.input_width);
@@ -135,24 +131,22 @@ void conv_general(
       for(T_UINT ki = 0; ki < p.kernel_height; ki++)
       {
         T_INT row = (wi * p.stride_along_height) - p.padding + ki;
-        inside_row = (row >= 0 && row < (T_INT) p.input_height);
-
+	if (row < 0 || row >= p.input_height) { current_kernel_index += p.kernel_width * p.kernel_depth; continue; }
+				   
         for(T_UINT kj = 0; kj < p.kernel_width; kj++)
         {
           T_INT col = (wj * p.stride_along_width)  - p.padding + kj;
-          inside_col = (col >= 0 && col < (T_INT) p.input_width);
+	  if (col < 0 || col >= p.input_width) { current_kernel_index += p.kernel_depth; continue; }
 
           for(T_UINT kz = 0; kz < p.kernel_depth; kz++)
           {
-            if (inside_row && inside_col) {
-              unsigned in_idx = row * (p.input_width * p.kernel_depth) + col * (p.kernel_depth) + kz;
-              unsigned k_idx = current_kernel_index + kernel_offset;
+	    unsigned in_idx = row * (p.input_width * p.kernel_depth) + col * (p.kernel_depth) + kz;
+	    unsigned k_idx = current_kernel_index + kernel_offset;	    
 
-              T in_data = input[in_idx];
-              T k_data = kernels[k_idx];
-
-              out += in_data * k_data;
-            }
+	    T in_data = input[in_idx];
+	    T k_data = kernels[k_idx];
+	    
+	    out += in_data * k_data;
             current_kernel_index++;
           }
         }
@@ -173,14 +167,13 @@ void convolution(
 {
   // use special implementation for 1x1 conv
   if (p.kernel_height == 1 && p.kernel_width == 1 && p.padding == 0) {
-    int kernels_size = p.kernel_height * p.kernel_width * p.kernel_depth * p.output_channels;
     conv1x1_kn2row(input, kernels, output, p);
     return;
   } else if (p.kernel_height == 3 && p.kernel_width == 3 && p.padding == 1) {
     int kernels_size = p.kernel_height * p.kernel_width * p.kernel_depth * p.output_channels;
-    T* kernels_hwoi = new T[kernels_size]();
+    T* kernels_hwoi = new T[kernels_size];
     ohwi_to_hwoi(kernels, kernels_hwoi, p);
-    conv3x3_kn2row(input, kernels_hwoi, output, p);
+    conv3x3_kn2row(input, kernels_hwoi, output, p);      
     delete[] kernels_hwoi;
     return;
   }