improve loop optimization.