@tkng 2018-12-10
For this kind of performance improvement, please share the result of benchmark.

Also, plesae include the procedure to reproduce the benchmark result, so that reviewer can check the result.