@lm-konda 2018-12-10
performance

before
```
Convolution,168176,2093,  sum:170.269ms
TotalRunTime,302676,  sum:302.676ms
```
after
```
Convolution,166709,2091,  sum:168.8ms
TotalRunTime,301169,  sum:301.169ms
```