@tkng 2018-12-10
Sorry, my words were not enough. Please include the following information.

- network you used
- dataset you used
- training configuration you used
- hardware you used
- how many times you run the benchmark (I want to know that whether the improvement is statistically significant.)

These're what I intended by the word "procedure to reproduce the benchmark result".