@lm-konda 2018-12-11
I am almost always using the configuration below.

```
network
　LmnetV1Quantize
dataset you used
　Caltech101
training configurattion
　epoch is 2. other is default.
hardware
　DE10-nano
how many times you run the benchmark
　As much as average average value is known (around 10 or more)
```