@tkng 2018-12-12
> As much as average average value is known (around 10 or more)

Please tell me the average and variance, since the effectiveness of this change is not obvious for me, "speed up to 168.8ms from 170.269ms" looks that just a fluctuation.