@lm-konda 2018-12-12
@tkng 
You are right.
There was no significant performance improvement when remeasurement was done. 
The reason is probably because I have cut down the necessary memset.
Sorry. I withdraw this PR.